+++ app/Http/Controllers/BreedingRecordController.php
@@ -4,34 +4,72 @@ namespace App\Http\Controllers;
 
 use App\Enums\AnimalHeatStatusEnum;
 use App\Http\Requests\BreedingRequest;
-use App\Models\Models\Animal;
-use App\Models\Models\BreedingRecord;
+use App\Models\Animal;
+use App\Models\BreedingRecord;
+use App\Models\User;
+use App\Repositories\BreedingFiltersRepository;
 use App\Services\ExceptionHandler;
 use App\Services\PredictBirthService;
+use Exception;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Http\Request;
 
 class BreedingRecordController extends Controller
 {
     protected $predictBirthService;
     protected $exceptions;
+    protected $filterRepository;
 
-    public function __construct(PredictBirthService $predictBirthService, ExceptionHandler $exceptions)
+    public function __construct(PredictBirthService $predictBirthService, ExceptionHandler $exceptions, BreedingFiltersRepository $breedingFiltersRepository)
     {
         $this->predictBirthService = $predictBirthService;
         $this->exceptions = $exceptions;
+        $this->filterRepository = $breedingFiltersRepository;
     }
 
-    public function index()
+    public function index(BreedingRecord $breedingRecord, Request $request)
     {
         try {
-            $matings = BreedingRecord::with(['female', 'male'])->get();
-            return response()->json($matings);
-        } catch (\Exception $exception) {
+            $this->authorize('viewAny', BreedingRecord::class);
+
+            $perPage = $request->input('per_page', 3);
+
+            $farmId = Auth::user()->farm_id;
+            $farmBreedings = BreedingRecord::whereHas('female', function ($query) use ($farmId) {
+                $query->where('farm_id', $farmId);
+            })->with(['female', 'male']);
+
+            $breedingsFiltered = $this->filterRepository->filter($farmBreedings, [
+                'occurrence_date' => $request->occurrence_date,
+                'coverage_date' => $request->coverage_date,
+                'date_birth_prediction' => $request->date_birth_prediction,
+                'date_birth' => $request->date_birth,
+                'status' => $request->status,
+                'cover_method' => $request->cover_method,
+                'female_id' => $request->female_id,
+                'male_id' => $request->male_id,
+            ]);
+
+            $paginatedBreedings = $breedingsFiltered->paginate($perPage, ['*'], 'page', $request->input('current_page', 1));
+
+            return response()->json([
+                'data' => $paginatedBreedings->items(),
+                'pagination' => [
+                    'total' => $paginatedBreedings->total(),
+                    'per_page' => $paginatedBreedings->perPage(),
+                    'current_page' => $paginatedBreedings->currentPage(),
+                    'last_page' => $paginatedBreedings->lastPage(),
+                ]
+            ]);
+        } catch (Exception $exception) {
             return $this->exceptions->getExceptions($exception);
         }
     }
 
     public function store(BreedingRequest $request)
     {
+        $this->authorize('create', BreedingRecord::class);
+
         $data = $request->validated();
         if (!array_key_exists('date_birth', $data) || ($data['status'] === AnimalHeatStatusEnum::PENDING)) {
             $data = $this->predictBirthService->birthPrediction($data);
@@ -39,21 +77,31 @@ class BreedingRecordController extends Controller
         $data = $this->predictBirthService->dateHandling($data);
 
         $breedingRecord = BreedingRecord::create($data);
-        return response()->json($breedingRecord, 201);
+        return response()->json(['breeding' => $breedingRecord], 201);
     }
 
-    public function show(BreedingRecord $breedingRecord)
+    public function show(BreedingRecord $model, $id)
     {
         try {
-            return response()->json($breedingRecord);
-        } catch (\Exception $exception) {
+            $breedingRecord = $model->find($id);
+            $this->authorize('view', $breedingRecord);
+
+            if (Auth::user()->farm_id === $breedingRecord->female->farm_id) {
+                return response()->json(['breeding' => $breedingRecord]);
+            }
+            return response()->json(['message' => 'Unauthorized'], 403);
+
+        } catch (Exception $exception) {
             return $this->exceptions->getExceptions($exception);
         }
     }
 
-    public function update(BreedingRequest $request, BreedingRecord $BreedingRecord)
+    public function update(BreedingRequest $request, BreedingRecord $model, $id)
     {
         try {
+            $breedingRecord = $model->find($id);
+            $this->authorize('update', $breedingRecord);
+
             $data = $request->validated();
 
             if (!array_key_exists('date_birth', $data) || ($data['status'] === AnimalHeatStatusEnum::PENDING)) {
@@ -64,18 +112,18 @@ class BreedingRecordController extends Controller
                 $data['status'] = AnimalHeatStatusEnum::SUCCESS;
             }
 
-            $BreedingRecord->update($data);
+            $breedingRecord->update($data);
 
             $female = Animal::find($data['female_id']);
-            $BreedingRecord->female()->associate($female);
+            $breedingRecord->female()->associate($female);
 
             if ($data['male_id']) {
                 $male = Animal::find($data['male_id']);
-                $BreedingRecord->male()->associate($male);
+                $breedingRecord->male()->associate($male);
             }
 
-            return response()->json(['success' => true, 'data' => $BreedingRecord->get()->first()], 200);
-        } catch (\Exception $exception) {
+            return response()->json(['success' => true, 'breeding' => $breedingRecord->get()->first()], 200);
+        } catch (Exception $exception) {
             return $this->exceptions->getExceptions($exception);
         }
     }
@@ -83,9 +131,11 @@ class BreedingRecordController extends Controller
     public function destroy(BreedingRecord $breedingRecord)
     {
         try {
+            $this->authorize('delete', $breedingRecord);
+
             $breedingRecord->delete();
             return response()->json(null, 204);
-        } catch (\Exception $exception) {
+        } catch (Exception $exception) {
             return $this->exceptions->getExceptions($exception);
         }
     }
